requirements.txt
langchain_groq
langchain-community 
langchain
python-dotenv
fastapi 
uvicorn 
streamlit 
pytest